+++ docs/changes.rst
@@ -38,6 +38,7 @@ v2.2.0 (in development)
 - Update: ``antics`` after feed change.
 - Update: ``beyondthetree`` is no longer published.
 - Update: ``chainsawsuit`` after feed change.
+- Update: ``choppingblock`` is no longer published.
 - Update: ``dilbertvg`` after move to new site.
 - Update: ``eatthattoast`` to not throw exception when the site is broken.
 - Update: ``evilinc`` to be more robust.